added the rest of the subrouters
